+++ lib/feature/repository/yaml_repository.rb
@@ -1,7 +1,7 @@
 module Feature
   module Repository
     ##
-    # YamlRepository for active and inactive feature list
+    # YamlRepository for active and inactive features
     # The yaml config file should look like this:
     #
     #     features:
@@ -40,27 +40,6 @@ module Feature
         features.sort.map(&:to_sym)
       end
       private :get_active_features_from_file
-
-      ##
-      # Returns list of inactive features
-      #
-      #   @return   Array<Symbol>
-      #
-      def inactive_features
-        get_inactive_features_from_file
-      end
-
-      ##
-      # Return a list of inactive features read from the config file
-      #
-      #   @return   Array<Symbol>
-      #
-      def get_inactive_features_from_file
-        features_hash = read_and_parse_file_data
-        features = features_hash.keys.select { |feature_key| !features_hash[feature_key] }
-        features.sort.map(&:to_sym)
-      end
-      private :get_inactive_features_from_file
 
       ##
       # Read and parses the feature config file